+++ src/variety/yajilin.js
@@ -5,13 +5,12 @@
 	if(typeof module==='object' && module.exports){module.exports = [pidlist, classbase];}
 	else{ pzpr.classmgr.makeCustom(pidlist, classbase);}
 }(
-['yajilin'], {
+['yajilin', 'yajilin-regions'], {
 //---------------------------------------------------------
 // マウス入力系
 MouseEvent:{
 	RBShadeCell : true,
 	use     : true,
-	inputModes : {edit:['number','direc','clear','info-line'],play:['line','peke','shade','unshade','info-line','completion']},
 	mouseinput_auto : function(){
 		if(this.puzzle.playmode){
 			if(this.mousestart || this.mousemove){
@@ -20,13 +19,20 @@ MouseEvent:{
 			}
 			else if(this.mouseend && this.notInputted()){
 				var cell = this.getcell();
-				if(!this.firstCell.isnull&&cell!==this.firstCell){ return;}
-				if(!cell.isnull && cell.isNum()){ this.inputqcmp();}
-				else                            { this.inputcell();}
+				if(!this.firstCell.isnull && cell!==this.firstCell){ return;}
+				if(!cell.isnull && cell.isNum() && this.pid==='yajilin'){
+					this.inputqcmp();
+				}
+				else{
+					this.inputcell();
+				}
 			}
 		}
 		else if(this.puzzle.editmode){
-			if(this.mousestart || this.mousemove){ this.inputdirec();}
+			if(this.mousestart || this.mousemove){
+				if(this.pid==='yajilin')             { this.inputdirec();}
+				else if(this.pid==='yajilin-regions'){ this.inputborder();}
+			}
 			else if(this.mouseend && this.notInputted()){ this.inputqnum();}
 		}
 	},
@@ -39,6 +45,12 @@ MouseEvent:{
 
 		this.mousereset();
 	}
+},
+"MouseEvent@yajilin":{
+	inputModes : {edit:['number','direc','clear','info-line'],play:['line','peke','shade','unshade','info-line','completion']}
+},
+"MouseEvent@yajilin-regions":{
+	inputModes : {edit:['number','border','clear','info-line'],play:['line','peke','shade','unshade','info-line']}
 },
 
 //---------------------------------------------------------
@@ -59,15 +71,26 @@ KeyEvent:{
 //---------------------------------------------------------
 // 盤面管理系
 Cell:{
+	// don't draw  dots under lines
+	isDot : function(){
+		return this.lcnt===0&&this.qsub===1;
+	},
+
+	knowEmpty : function(){
+		if(this.qsub!==0){ return true;}
+		if(this.lcnt>0){ return true;}
+		var shadedNbrs = this.countDir4Cell(function(cell){
+			return (cell.isShade() > 0);
+		});
+		return (shadedNbrs>0);
+	}
+},
+"Cell@yajilin":{
 	minnum : 0,
 	maxnum : function(){ return Math.max((this.board.cols+1)>>1,(this.board.rows+1)>>1);},
 
 	// 線を引かせたくないので上書き
 	noLP : function(dir){ return (this.isShade() || this.isNum());},
-
-	isDot : function(){
-		return this.lcnt===0&&this.qsub===1;
-	},
 
 	allowShade : function(){
 		return this.qnum===-1&&this.lcnt===0;
@@ -104,14 +127,6 @@ Cell:{
 		if(!clist){ return -1;}
 		return (clist.filter(function(cell){ return cell.isShade();}).length);
 	},
-	knowEmpty : function(){
-		if(this.qsub!==0){ return true;}
-		if(this.lcnt>0){ return true;}
-		var shadedNbrs = this.countDir4Cell(function(cell){
-			return (cell.isShade() > 0);
-		});
-		return (shadedNbrs>0);
-	},
 	countUndecided : function(clist){
 		if(!clist){ return -1;}
 		return (clist.filter(function(cell){
@@ -122,6 +137,7 @@ Cell:{
 		}).length);
 	},
 
+	// trigger redraw for autocompletion
 	posthook : {
 		qsub : function(){
 			var cells=[this];
@@ -139,9 +155,23 @@ Cell:{
 		}
 	}
 },
+"Cell@yajilin-regions":{
+	minnum : 0,
+	maxnum : function(){ return this.room.clist.length;},
+
+	noLP : function(dir){ return this.isShade();},
+
+	allowShade : function(){
+		return this.lcnt===0;
+	},
+	allowUnshade : function(){
+		return this.lcnt===0;
+	}
+},
 Border:{
-	enableLineNG : true,
-	
+	enableLineNG : true
+},
+"Border@yajilin":{
 	isBorder : function(){
 		return (this.sidecell[0].qnum===-1)!==(this.sidecell[1].qnum===-1);
 	},
@@ -156,7 +186,8 @@ Border:{
 },
 Board:{
 	hasborder : 1,
-
+},
+"Board@yajilin":{
 	redrawAffected : function(cells){
 		var minx=this.maxbx, maxx=this.minbx, miny=this.maxby, maxy=this.minby;
 		for(var i=0;i<cells.length;i++){
@@ -181,12 +212,15 @@ Board:{
 		}
 	}
 },
-BoardExec:{
+"BoardExec@yajilin":{
 	adjustBoardData : function(key,d){
 		this.adjustNumberArrow(key,d);
 	}
 },
-
+"AreaRoomGraph@yajilin-regions":{
+	enabled : true,
+	hastop : true
+},
 LineGraph:{
 	enabled : true
 },
@@ -196,25 +230,31 @@ LineGraph:{
 Graphic:{
 	irowake : true,
 
-	qcmpcolor  : "rgb(127,127,127)",
-	autocmp : 'number',
-
 	paint : function(){
 		this.drawBGCells();
+		if(this.pid==='yajilin-regions'){ this.drawShadedCells();}
 		this.drawDotCells();
 		this.drawGrid();
 		
+
 		this.drawBorders();
 
-		this.drawArrowNumbers();
+		if(this.pid==='yajilin'){ this.drawArrowNumbers();}
 
 		this.drawLines();
+
+		if(this.pid==='yajilin-regions'){ this.drawQuesNumbers();}
+
 		this.drawPekes();
 
 		this.drawChassis();
 
 		this.drawTarget();
-	},
+	}
+},
+"Graphic@yajilin":{
+	qcmpcolor  : "rgb(127,127,127)",
+	autocmp : 'number',
 
 	getBGCellColor : function(cell){
 		var info = cell.error || cell.qinfo;
@@ -240,7 +280,7 @@ Graphic:{
 
 //---------------------------------------------------------
 // URLエンコード/デコード処理
-Encode:{
+"Encode@yajilin":{
 	decodePzpr : function(type){
 		this.decodeArrowNumber16();
 		
@@ -259,8 +299,18 @@ Encode:{
 		this.fio.encodeCellDirecQnum_kanpen(true);
 	}
 },
+"Encode@yajilin-regions":{
+	decodePzpr : function(type){
+		this.decodeBorder();
+		this.decodeRoomNumber16();
+	},
+	encodePzpr : function(type){
+		this.encodeBorder();
+		this.encodeRoomNumber16();
+	}
+},
 //---------------------------------------------------------
-FileIO:{
+"FileIO@yajilin":{
 	decodeData : function(){
 		this.decodeCellDirecQnum();
 		this.decodeCellAns();
@@ -356,7 +406,20 @@ FileIO:{
 		});
 	}
 },
-
+"FileIO@yajilin-regions":{
+	decodeData : function(){
+		this.decodeAreaRoom();
+		this.decodeCellQnum();
+		this.decodeCellAns();
+		this.decodeBorderLine();
+	},
+	encodeData : function(){
+		this.encodeAreaRoom();
+		this.encodeCellQnum();
+		this.encodeCellAns();
+		this.encodeBorderLine();
+	}
+},
 //---------------------------------------------------------
 // 正解判定処理実行部
 AnsCheck:{
@@ -366,14 +429,19 @@ AnsCheck:{
 		"checkLineOnShadeCell",
 		"checkAdjacentShadeCell",
 		"checkDeadendLine+",
-		"checkArrowNumber",
+		"checkArrowNumber@yajilin",
+		"checkShadeCellCount@yajilin-regions",
 		"checkOneLoop",
-		"checkEmptyCell_yajilin+",
-		"checkNumberHasArrow",
+		"checkEmptyCell_yajilin+@yajilin",
+		"checkEmptyCell_regions+@yajilin-regions",
+		"checkNumberHasArrow@yajilin",
 	],
 
 	checkEmptyCell_yajilin : function(){
 		this.checkAllCell(function(cell){ return (cell.lcnt===0 && !cell.isShade() && cell.noNum());}, "ceEmpty");
+	},
+	checkEmptyCell_regions : function(){
+		this.checkAllCell(function(cell){ return (cell.lcnt===0 && !cell.isShade());}, "ceEmpty");
 	},
 
 	checkArrowNumber : function(){